Footer links added & placeholders removed for content
If applied, this commit will have updated the placeholder logos in the footer of about.html to include real URLs to Github and LinkedIn pages on contact.html file.

If applied, this commit will have updated the content of the description, no formatting changes, only text within table and h1.